接口添加身份权限验证

+++ backend/app/Http/Controllers/Api/V1/ActivityController.php
@@ -10,8 +10,15 @@ use App\Model\Card;
 use App\Model\Shop;
 use App\Utils\ResponseMessage;
 use Illuminate\Http\Request;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class ActivityController extends ApiController {
+
+    private $oneself;
+
+    public function __construct() {
+        $this->oneself = JWTAuth::parseToken()->authenticate();
+    }
 
     /**
      * 获取所有活动列表
@@ -19,6 +26,10 @@ class ActivityController extends ApiController {
      * @return mixed
      */
     public function list() {
+        if ($notAdmin = $this->isAdmin($this->oneself)) {
+            return $notAdmin;
+        }
+
         return $this->success(
             Activity::all()
                 ->map(function ($item) {
@@ -65,6 +76,10 @@ class ActivityController extends ApiController {
      * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response|mixed
      */
     public function store(StoreActivityRequest $request) {
+        if ($notAdmin = $this->isAdmin($this->oneself)) {
+            return $notAdmin;
+        }
+
         if ($request->has("reply_keyword")
             && Activity::where("reply_keyword", $request->get("reply_keyword"))->get()->isNotEmpty()) {
             return $this->badRequest(NULL, ResponseMessage::$message[400019]);
@@ -88,7 +103,7 @@ class ActivityController extends ApiController {
             $activity->state = $request->get("state");
         }
 
-        $this->save_model($activity);
+        $this->saveModel($activity);
 
         if ($request->has("card_model_id_list")) {
             $activity->cards()->attach($request->get("card_model_id_list"));
@@ -106,6 +121,10 @@ class ActivityController extends ApiController {
      * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
      */
     public function update(UpdateActivityRequest $request, $id) {
+        if ($notAdmin = $this->isAdmin($this->oneself)) {
+            return $notAdmin;
+        }
+
         $activity = Activity::find($id);
 
         if ( ! $activity) {
@@ -146,7 +165,7 @@ class ActivityController extends ApiController {
             $request->card_model_id_list = $request->get("card_model_id_list") ? $request->get("card_model_id_list") : [];
         }
 
-        $this->save_model($activity);
+        $this->saveModel($activity);
 
         if ($request->has("card_model_id_list")) {
             $activity->cards()->detach();
@@ -160,8 +179,14 @@ class ActivityController extends ApiController {
      * 删除活动
      *
      * @param $id
+     *
+     * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
      */
     public function remove($id) {
+        if ($notAdmin = $this->isAdmin($this->oneself)) {
+            return $notAdmin;
+        }
+
         $activity = Activity::find($id);
         $activity->shops()->detach();
         $activity->cards()->detach();
